Windows: Build load.go
