+++ graph/load.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build linux windows
 
 package graph
