+++ graph/tags_unit_test.go
@@ -181,24 +181,6 @@ func TestLookupImage(t *testing.T) {
 	}
 }
 
-func TestValidTagName(t *testing.T) {
-	validTags := []string{"9", "foo", "foo-test", "bar.baz.boo"}
-	for _, tag := range validTags {
-		if err := ValidateTagName(tag); err != nil {
-			t.Errorf("'%s' should've been a valid tag", tag)
-		}
-	}
-}
-
-func TestInvalidTagName(t *testing.T) {
-	validTags := []string{"-9", ".foo", "-test", ".", "-"}
-	for _, tag := range validTags {
-		if err := ValidateTagName(tag); err == nil {
-			t.Errorf("'%s' shouldn't have been a valid tag", tag)
-		}
-	}
-}
-
 func TestValidateDigest(t *testing.T) {
 	tests := []struct {
 		input       string
